Shrink the generated validator: minify + drop zod from its bundle
The esbuild-bundled validate-bundle.js pulled in all of zod unminified
(378 lines/12KB -> 15,049 lines/550KB), because validate.ts imported
ProjectBundleSchema (for parseBundle) at module scope, which forces
the whole zod-based bundle.ts/enums.ts/playlist.ts construction to run
even though validateBundle() itself never touches zod at runtime.

- Move parseBundle into its own parse.ts (the only zod-dependent piece)
- Move the plain ID arrays into a new zod-free ids.ts, imported by both
  enums.ts (for the zod-wrapped schemas) and validate.ts (semantic
  rules only) — @arkaik/schema's public exports are unchanged
- Enable esbuild minification for the CLI build

Result: validate-bundle.js is now 18 lines / 9KB, genuinely
zero-dependency, same CLI contract and report as before.
