+++ packages/schema/src/index.ts
@@ -2,3 +2,4 @@ export * from "./enums";
 export * from "./playlist";
 export * from "./bundle";
 export * from "./validate";
+export * from "./parse";
